bake-decouple: move dev-server inspector agent into bake
