+++ src/codegen/generate-host-exports.ts
@@ -503,7 +503,7 @@ const importCandidates: Array<[string, string]> = [
   ["bun_jsc", "JSPromise"],
   ["bun_jsc", "ZigStackFrame"],
   ["bun_jsc::virtual_machine", "VirtualMachine"],
-  ["bun_jsc::debugger", "InspectorBunFrontendDevServerAgentHandle"],
+  ["crate::bake::dev_server::inspector_agent", "InspectorBunFrontendDevServerAgentHandle"],
   ["bun_jsc::debugger", "LifecycleHandle"],
   ["bun_jsc::debugger", "TestReporterHandle"],
 ];
